+++ Server/config.py
@@ -1,0 +1,3 @@
+class FlaskConfig:
+    SECRET_KEY = 'TEMP_SECRET_KEY'
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///site.db'
