+++ Server/templates/classroom.html
@@ -5,7 +5,7 @@
 				<h3>Class: {{ current_class.name }}</h3>
 			</div>
 			<div class="col-1">
-				<a role="button" class="btn btn-dark" href="{{ url_for('home') }}"><span class="fas fa-long-arrow-alt-left"></span></a>
+				<a role="button" class="btn btn-dark" href="{{ url_for('classrooms.home') }}"><span class="fas fa-long-arrow-alt-left"></span></a>
 			</div>
 		</div>
 		<div class="row">
